Move Outpost2DLL code to game directory
