+++ Outpost2DLL.h
@@ -10,20 +10,20 @@
 
 // Include all the header files which contain
 // defintions of exports from Outpost2.exe
-#include "MapIdEnum.h"
-#include "Enums.h"
-#include "Structs.h"
-#include "Player.h"
-#include "Unit.h"
-#include "UnitBlock.h"
-#include "ScStub.h"
-#include "Groups.h"
-#include "Trigger.h"
-#include "Enumerators.h"
-#include "TethysGame.h"
-#include "GameMap.h"
-#include "NonExportedEnums.h"
-#include "Globals.h"
+#include "game/MapIdEnum.h"
+#include "game/Enums.h"
+#include "game/Structs.h"
+#include "game/Player.h"
+#include "game/Unit.h"
+#include "game/UnitBlock.h"
+#include "game/ScStub.h"
+#include "game/Groups.h"
+#include "game/Trigger.h"
+#include "game/Enumerators.h"
+#include "game/TethysGame.h"
+#include "game/GameMap.h"
+#include "game/NonExportedEnums.h"
+#include "game/Globals.h"
 // Include the header file which contains definitions
 // of required exports from the level DLL
-#include "RequiredExports.h"
+#include "game/RequiredExports.h"
