+++ game/Outpost2DLL.cpp
@@ -1,4 +1,4 @@
 // This is a dummy source file for what is effectively a header only library.
 // Adding it to the project allows the project to build, which runs all
 // included headers through the compiler, checking for syntax errors.
-#include "Outpost2DLL.h"
+#include "../Outpost2DLL.h"
